% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_clean.R
\name{clean_coord}
\alias{clean_coord}
\title{Clean coordinates}
\usage{
clean_coord(coord, na_char = "^\\\\s*$")
}
\arguments{
\item{coord}{coordinates vector}

\item{na_char}{Regular expression to matcxh and replace with NA.
Defaults to all blank characters (or empty characters)}
}
\value{
the coordinates with only numbers and point as a decimal
separator (instead of comma).
}
\description{
Clean coordinates
}
