% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_clean.R
\name{clean_count}
\alias{clean_count}
\title{Chean count values}
\usage{
clean_count(vec, na_char = "^\\\\s*$")
}
\arguments{
\item{vec}{vector of counts}

\item{na_char}{values to replace by \code{NA}}
}
\value{
The count values cleaned and coerced to numeric.
Characters matching \code{na_char} are converted to \code{NA} and
"x" and "X" are converted to 1.
}
\description{
Clean count values
}
