% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_clean.R
\name{coerce_to_na}
\alias{coerce_to_na}
\title{Coerce to NA}
\usage{
coerce_to_na(vec, na_char = "^\\\\s*$")
}
\arguments{
\item{vec}{the vector to modify}

\item{na_char}{the characters t convert to NA: defaults to blank characters}
}
\value{
the vector where values matching \code{na_char} have been converted to \code{NA}
}
\description{
Convert values of a vector to \code{NA}
}
