% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_clean.R
\name{coerced_to_na}
\alias{coerced_to_na}
\title{Check NA values}
\usage{
coerced_to_na(before, after, return = FALSE)
}
\arguments{
\item{before}{The vector before conversion}

\item{after}{The vector after conversion}

\item{return}{Return the detailed results?}
}
\value{
If \code{return = FALSE} (default), prints a message to warn for
introduced NAs.
Else, returns the indices of values with NA and the initial values
converted to NA.
}
\description{
Check which values have been coeced to NA.
}
