% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{colnames_DB}
\alias{colnames_DB}
\title{Get DB column names}
\usage{
colnames_DB(connexion, table, rm_ID = TRUE)
}
\arguments{
\item{connexion}{connexion to the table (class \code{PqConnection})}

\item{table}{table to retrieve names from}

\item{rm_ID}{Remove ID column? (columns that end with ID)}
}
\value{
The table column names
}
\description{
Get DB column names
}
