% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{df_all_cols}
\alias{df_all_cols}
\title{Get all columns of a dataframe}
\usage{
df_all_cols(df, cols)
}
\arguments{
\item{df}{the original df}

\item{cols}{the columns we want in the final df}
}
\value{
a dataframe where all colnames in \code{cols} are present:
if the column was in the original table, values are conserved,
else they are filled with \code{NA}s.
}
\description{
Get all columns, even if they don't exist
}
