% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragondb-package.R
\docType{package}
\name{dragondb-package}
\alias{dragondb}
\alias{dragondb-package}
\title{dragondb: Format and Insert Data in Dragon Database}
\description{
Package to prepare data for integration in the Dragon database.
}
\author{
\strong{Maintainer}: Lisa Nicvert \email{lisa.nicvert@fondationbiodiversite.fr} (\href{https://orcid.org/0009-0006-5763-0865}{ORCID})

}
\keyword{internal}
