% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_clean.R
\name{recode_counts}
\alias{recode_counts}
\title{Recode counts values}
\usage{
recode_counts(
  d,
  key,
  variable.name = "lifeStage",
  value.name = "individualCount"
)
}
\arguments{
\item{d}{data.table}

\item{key}{key to get the values that each column encodes (these values
replace the old column names in column \code{variable.name})}

\item{variable.name}{The name of the column holding variables}

\item{value.name}{The name of the column holding values}
}
\value{
The modofied data.table
}
\description{
This function recodes counts that are encoded in different columns
to a single column, and a second column indicating individual type
(under the hood, it melts the data.table to long)
}
