% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_clean.R
\name{rename_cols}
\alias{rename_cols}
\title{Rename columns}
\usage{
rename_cols(key, dtable, nam)
}
\arguments{
\item{key}{a dataframe containing the correspondence between the new
(in column \code{Standard}) and old (in column \code{nam}) column names.}

\item{dtable}{A datatable for which columns must be renamed}

\item{nam}{Name of the column of \code{key} in which to search old names.}
}
\value{
No value, sets the column names of \code{dtable}
}
\description{
Rename columns
}
