% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{replace_NA}
\alias{replace_NA}
\title{Replace \code{NA} values}
\usage{
replace_NA(dt, SDcols, placeholder = "__NA__", rev = FALSE)
}
\arguments{
\item{dt}{the data.table}

\item{SDcols}{the columns in which to replace values}

\item{placeholder}{the placeholder}

\item{rev}{reverse (i.e. replace placeholder back with \code{NA}).}
}
\value{
the data.table where NA values are replaced with \code{placeholder}
(or the reverse if \code{rev} is \code{TRUE})
}
\description{
Replace \code{NA} values in a data.table with a placeholder
}
