% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{rm_all_na}
\alias{rm_all_na}
\title{Remove rows with all NAs}
\usage{
rm_all_na(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
the same dataframe where rows containing only \code{NA}s
have been discarded.
}
\description{
Remove rows with all NAs
}
