% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_clean.R
\name{rm_spaces}
\alias{rm_spaces}
\title{Remove leading and trailing spaces}
\usage{
rm_spaces(vec)
}
\arguments{
\item{vec}{vector in which spaces should be removed}
}
\value{
values in the vector without leading or trailing spaces.
Multiple instances of blank characters are removed.
}
\description{
Remove leading and trailing spaces
}
